requirements.txt
# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0
chardet>=5.2.0

# Dataset Creation
wikipedia-api>=0.8.1

# Embeddings
sentence-transformers>=2.3.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0

# Vector Database
chromadb>=0.5.0

# Visualization
umap-learn>=0.5.5
plotly>=5.18.0
matplotlib>=3.8.0

# Jupyter Widgets
ipywidgets>=8.1.0
tqdm>=4.66.0
jupyter>=1.0.0

# LLM - Gemini
google-generativeai>=0.3.0

# Utilities
pandas>=2.2.0
python-dotenv>=1.0.0
